+++ apps/anime-service/src/modules/dates/items/dates-item.ts
@@ -1,0 +1,9 @@
+export abstract class DatesItem {
+  public currentDate;
+
+  constructor() {
+    this.currentDate = new Date().toISOString();
+  }
+
+  abstract getDates(startDate: Date, endDate: Date): void;
+}
