+++ apps/anime-service/src/modules/dates/items/future-dates.ts
@@ -1,0 +1,7 @@
+import { DatesItem } from './dates-item';
+
+export class FutureDates extends DatesItem {
+  getDates(startDate: Date, endDate: Date) {
+    return { startDate: null, endDate: null };
+  }
+}
