+++ apps/anime-service/src/modules/dates/items/past-dates.ts
@@ -1,0 +1,7 @@
+import { DatesItem } from './dates-item';
+
+export class PastDates extends DatesItem {
+  getDates(startDate: Date, endDate: Date) {
+    return { startDate, endDate: endDate ?? this.currentDate };
+  }
+}
