+++ apps/anime-service/src/modules/dates/items/present-dates.ts
@@ -1,0 +1,7 @@
+import { DatesItem } from './dates-item';
+
+export class PresentDates extends DatesItem {
+  getDates(startDate: Date, endDate: Date) {
+    return { startDate: startDate ?? this.currentDate, endDate: null };
+  }
+}
